[IMP] account{edi_ubl_cii,peppol}: Self-Billed invoice import/export
Self-billed invoices are invoices which are created by the customer on
behalf of the vendor. These can be used in certain business flows,
typically when the customer needs to create an invoice internally for
accounting purposes.

Peppol provides the ability for customers to send self-billed invoices
to the vendor over the Peppol network. This way the vendor can verify
whether the invoice is correct.

See the Peppol documentation at
https://docs.peppol.eu/poacc/self-billing/3.0/bis-sb/

This commit provides
(1) the ability to export vendor bills as self-billed invoices
conformant to the BIS3 standard
(2) the ability to import BIS3 self-billed invoices as customer invoices
(3) the ability to send vendor bills as self-billed invoices over Peppol
(4) the ability to receive self-billed invoices from Peppol in a
specified Sales journal.

Configuration
In the Peppol advanced configuration wizard, we add a boolean field to
allow sending vendor bills as self-billed invoices, and a Many2one to
a Sales `account.journal` to select the reception journal for
self-billed invoices.

We also add the self-billed invoice Peppol document type IDs in the list
of documents which the user can select or deselect to allow/prevent
reception of those document types.

task-4774669

closes odoo/odoo#222658

Related: odoo/enterprise#92205
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>